Fix clang -std=c++17 build
Note: still building Ice C++11 with -std=c++11.

+++ cpp/src/Ice/InputStream.cpp
@@ -481,7 +481,7 @@ struct ReadBoolHelper
 {
     static bool* read(pair<const bool*, const bool*>& v, Int sz, InputStream::Container::iterator& i)
     {
-        bool* array = new bool[sz];
+        bool* array = new bool[static_cast<size_t>(sz)];
         for(int idx = 0; idx < sz; ++idx)
         {
             array[idx] = static_cast<bool>(*(i + idx));
